+++ tests/unit/test_inspection.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+
+from expects import expect
+
+from booby import models, fields, errors
+from booby.inspection import inspect
+
+
+class TestInspect(object):
+    def test_instance_should_have_fields_attr_with_the_dict_of_fields(self):
+        model = inspect(User())
+
+        expect(model.fields).to.have.keys(name=User.name, email=User.email)
+
+    def test_class_should_have_fields_attr_with_the_dict_of_fields(self):
+        model = inspect(User)
+
+        expect(model.fields).to.have.keys(name=User.name, email=User.email)
+
+    def test_non_model_object_should_raise_inspect_error(self):
+        expect(lambda: inspect(object)).to.raise_error(errors.InspectError)
+
+
+class User(models.Model):
+    name = fields.String()
+    email = fields.String()